third_party/blink/web_tests/fast/scrolling/events/overscroll-event-fired-to-scrolled-element.html

<!DOCTYPE HTML>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/gesture-util.js"></script>
<script src="../../../virtual/percent-based-scrolling/resources/percent-based-util.js"></script>
<style>
#scrollableDiv {
  width: 200px;
  height: 200px;
  overflow: scroll;
}

#innerDiv {
  width: 400px;
  height: 400px;
}
</style>

<body style="margin:0">
<div id="scrollableDiv">
  <div id="innerDiv">
  </div>
</div>
</body>

<script>
if (window.internals)
  internals.runtimeFlags.scrollEndEventsEnabled = true;

var scrolling_div = document.getElementById('scrollableDiv');
var rect = scrolling_div.getBoundingClientRect();
var x = rect.left + rect.width / 2;
var y = rect.top + rect.height / 2;

// SPEED_INSTANT doesn't work here because tests are timing out.
// This could actually be a bug in how the overscroll event is computed.
// Possible explanation on what happens is that "leftover" delta from a single
// event doesn't cause an overscroll event, only when the next event arrives
// and the full delta is overscroll do we dispatch the event.
// When the speed is INSTANT we'll put the entire scrolling delta into a single
// event so we don't generate an overscroll event in that case.
const pixelsPerSec = 2000;
const pixelsToScroll = 800;
const { x: pixelsToScrollX, y: pixelsToScrollY } = calculatePixelsToScroll(
  scrolling_div, pixelsToScroll, pixelsToScroll
);

var overscrolled_x_delta = 0;
var overscrolled_y_delta = 0;
function onOverscroll(event) {
  assert_false(event.cancelable);
  assert_false(event.bubbles);
  overscrolled_x_delta = event.deltaX;
  overscrolled_y_delta = event.deltaY;
}
scrolling_div.addEventListener("overscroll", onOverscroll);

async function test_overscroll_event_fire(source_device,
    precise_scrolling_delta) {
  scrolling_div.scrollTo(0, 0);
  overscrolled_x_delta = 0;
  overscrolled_y_delta = 0;
  await waitForCompositorCommit();

  // When PBS is enabled, the number of 'pixels to scroll' is significantly distorted
  // (~10k with current params), causing this test to take longer and time out
  // Adjusting speed by the distortion means the scroll anomations take the same
  // time whether PBS is on or off
  const scroll_speed_factor_x = pixelsToScrollX / pixelsToScroll;
  const scroll_speed_factor_y = pixelsToScrollY / pixelsToScroll;

  // Do a horizontal scroll and wait for overscroll event.
  await smoothScroll(
    pixelsToScrollX, x, y, source_device, "right", pixelsPerSec * scroll_speed_factor_x,
    precise_scrolling_delta
  );
  await waitFor(() => { return overscrolled_x_delta > 0; },
      'Scroller did not receive overscroll event after horizontal scroll.');
  await waitForAnimationEndTimeBased(() => { return scrolling_div.scrollLeft });
  assert_equals(scrolling_div.scrollWidth - scrolling_div.scrollLeft,
              scrolling_div.clientWidth);

  overscrolled_x_delta = 0;
  overscrolled_y_delta = 0;

  // Do a vertical scroll and wait for overscroll event.
  await smoothScroll(
    pixelsToScrollY, x, y, source_device, "down", pixelsPerSec * scroll_speed_factor_y,
    precise_scrolling_delta
  );
  await waitFor(() => { return overscrolled_y_delta > 0; },
      'Scroller did not receive overscroll event after vertical scroll.');
  await waitForAnimationEndTimeBased(() => { return scrolling_div.scrollTop });
  assert_equals(scrolling_div.scrollHeight - scrolling_div.scrollTop,
              scrolling_div.clientHeight);
}

promise_test (async (t) => {
  // Make sure that no overscroll event is sent to document.
  document.addEventListener("overscroll",
      t.unreached_func("Document got unexpected overscroll event."));
  return test_overscroll_event_fire(GestureSourceType.TOUCH_INPUT);
}, 'Tests that the scrolled element gets overscroll event after fully scrolling by touch.');

promise_test (async (t) => {
  // Make sure that no overscroll event is sent to document.
  document.addEventListener("overscroll",
      t.unreached_func("Document got unexpected overscroll event."));
  return test_overscroll_event_fire(GestureSourceType.MOUSE_INPUT,
      true /* precise_scrolling_delta */);
}, 'Tests that the scrolled element gets overscroll event after fully' +
   'scrolling by non-animating wheel scrolling.');

promise_test (async (t) => {
  // Make sure that no overscroll event is sent to document.
  document.addEventListener("overscroll",
      t.unreached_func("Document got unexpected overscroll event."));
  return test_overscroll_event_fire(GestureSourceType.MOUSE_INPUT,
      false /* precise_scrolling_delta */);
}, 'Tests that the scrolled element gets overscroll event after fully' +
   'scrolling by animated wheel scrolling (smooth-scrolling)');
</script>
